Move most source files into logical subpackages
In no way are these divisions set in stone, we can easily reorganize. The only change this causes is that we need to specify the specific subpackage, and explicitly import from other packages.
