+++ src/jerra/Main.java
@@ -2,6 +2,12 @@ package jerra;
 
 import java.util.Scanner;
 
+import jerra.room.Room;
+import jerra.room.TextRoom;
+
+import jerra.control.Controller;
+import jerra.control.TextController;
+
 // The main file
 public class Main {
 	public static void main(String[] args) {
