+++ src/jerra/control/Controller.java
@@ -1,4 +1,4 @@
-package jerra;
+package jerra.control;
 
 /**
  * Controller
