+++ src/jerra/core/Vector.java
@@ -1,4 +1,4 @@
-package jerra;
+package jerra.core;
 
 /**
  * Vector
