+++ src/jerra/entity/Bullet.java
@@ -1,4 +1,10 @@
-package jerra;
+package jerra.entity;
+
+import jerra.core.Rect;
+import jerra.core.Vector;
+
+import jerra.presence.Presence;
+import jerra.presence.DefaultPresence;
 
 /**
  * Bullet
