+++ src/jerra/entity/Entity.java
@@ -1,4 +1,11 @@
-package jerra;
+package jerra.entity;
+
+import jerra.api.Copyable;
+import jerra.api.Updatable;
+
+import jerra.core.Rect;
+
+import jerra.presence.Presence;
 
 /**
  * Entity
