+++ src/jerra/entity/Projectile.java
@@ -1,4 +1,6 @@
-package jerra;
+package jerra.entity;
+
+import jerra.presence.Presence;
 
 /**
  * Projectile
