+++ src/jerra/item/Caffeine.java
@@ -1,4 +1,4 @@
-package jerra;
+package jerra.item;
 
 /**
  * Caffeine
