+++ src/jerra/item/HalfAndHalf.java
@@ -1,4 +1,4 @@
-package jerra;
+package jerra.item;
 
 /**
  * HalfAndHalf
