+++ src/jerra/item/Inventory.java
@@ -1,4 +1,4 @@
-package jerra;
+package jerra.item;
 
 public interface Inventory {
 }
